samba-tool domain trust: allow using smb2 by default
Windows Server 1709 defaults to SMB2 and does not have SMB1 enabled.
When establishing trust, samba-tool does not specify SMB protocol
version and fail by default.

We have two ways to handle this situation:
 - switch 'client ipc min protocol' to 'SMB2' or later
 - explicitly specify 'smb2' to binding options and retry

Using first approach would be better in a longer term but right now Samba
defaults to 'CORE' in 'client min protocol' and changing it would deny
connecting to older SMB servers.

Another way is to explicitly specify a protocol in a binding string and
fall back to a default if connection fails. A similar approach is used
by the FreeIPA trust code for several years.

Signed-off-by: Alexander Bokovoy <ab@samba.org>