python/samba/tests: Fix decode error
This seems to be a side affect of using StringIO.StringIO in python2.
We need to use StringIO.StringIO (instead of cStringIO.StringIO)

Note: https://docs.python.org/2/library/stringio.html states

"The StringIO object can accept either Unicode or 8-bit strings, but
 mixing the two may take some care. If both are used, 8-bit strings
that cannot be interpreted as 7-bit ASCII (that use the 8th bit) will
cause a UnicodeError to be raised when getvalue() is called"

and also about cStringIO

"Unlike the StringIO module, this module is not able to accept Unicode
strings that cannot be encoded as plain ASCII strings.

We don't use cStringIO because now we do have more unicode (from porting efforts
to make the code py2/py3 compatible). For sure some of these unicode string
cannot be encoded as plain ASCII.

running test 'samba.tests.dcerpc.dnsserver' (with python2)  results in
UnicodeDecodeError: 'ascii' codec can't decode byte 0xff in position 26: ordinal not in range(128)

Change the native string (as it is in python2) that contains a non ascii character
to unicode. (in a plain python3 only world this would not be an issue)

squas

Signed-off-by: Noel Power <noel.power@suse.com>